+++ pwny/__init__.py
@@ -8,3 +8,4 @@ from pwnypack.codec import *
 from pwnypack.shell import *
 from pwnypack.rop import *
 from pwnypack.fmtstring import *
+from pwnypack.php import *